(#26) Moving classes between packages
