--- src/main/java/apps/amaralus/qa/platform/runtime/StepAction.java
@@ -1,7 +1,0 @@
-package apps.amaralus.qa.platform.runtime;
-
-@FunctionalInterface
-public interface StepAction {
-
-    void execute(TestContext testContext);
-}
