+++ src/main/java/apps/amaralus/qa/platform/runtime/execution/ExecutorServiceAware.java
@@ -1,4 +1,4 @@
-package apps.amaralus.qa.platform.runtime;
+package apps.amaralus.qa.platform.runtime.execution;
 
 import java.util.concurrent.ExecutorService;
