+++ src/main/java/apps/amaralus/qa/platform/runtime/execution/SimpleTask.java
@@ -1,4 +1,4 @@
-package apps.amaralus.qa.platform.runtime;
+package apps.amaralus.qa.platform.runtime.execution;
 
 import lombok.extern.slf4j.Slf4j;
 import org.jetbrains.annotations.Nullable;
