+++ src/main/java/apps/amaralus/qa/platform/runtime/execution/StepAction.java
@@ -1,0 +1,9 @@
+package apps.amaralus.qa.platform.runtime.execution;
+
+import apps.amaralus.qa.platform.runtime.TestContext;
+
+@FunctionalInterface
+public interface StepAction {
+
+    void execute(TestContext testContext);
+}
